+++ tests/pjit_test.py
@@ -50,7 +50,7 @@ from jax._src.sharding_impls import (
     SingleDeviceSharding, parse_flatten_op_sharding)
 import jax._src.pjit as pjit_lib
 from jax._src.pjit import pjit, pjit_p
-from jax._src import mesh
+from jax._src import mesh as mesh_lib
 from jax._src.interpreters import pxla
 from jax.interpreters import mlir
 from jax._src import xla_bridge
@@ -248,20 +248,20 @@ class PJitTest(jtu.BufferDonationTestCase):
   def testDifferentNestedMesh(self):
     with jtu.create_global_mesh((2, 1), ("x", "y")) as m1:
       with jtu.create_global_mesh((2, 2), ("a", "b")) as m2:
-        self.assertEqual(mesh.thread_resources.env.physical_mesh, m2)
-      self.assertEqual(mesh.thread_resources.env.physical_mesh, m1)
-    self.assertEqual(mesh.thread_resources.env.physical_mesh,
-                     mesh.EMPTY_ENV.physical_mesh)
+        self.assertEqual(mesh_lib.thread_resources.env.physical_mesh, m2)
+      self.assertEqual(mesh_lib.thread_resources.env.physical_mesh, m1)
+    self.assertEqual(mesh_lib.thread_resources.env.physical_mesh,
+                     mesh_lib.EMPTY_ENV.physical_mesh)
 
   def testSameNestedMesh(self):
     mesh = jtu.create_global_mesh((2, 1), ("a", "b"))
-    thread_resources = jax._src.mesh.thread_resources
+    thread_resources = mesh_lib.thread_resources
     with mesh as m1:
       with mesh as m2:
         self.assertEqual(thread_resources.env.physical_mesh, m2)
       self.assertEqual(thread_resources.env.physical_mesh, m1)
     self.assertEqual(thread_resources.env.physical_mesh,
-                     jax._src.mesh.EMPTY_ENV.physical_mesh)
+                     mesh_lib.EMPTY_ENV.physical_mesh)
 
   def testMeshDecorator(self):
     x = jnp.arange(8)